+++ src/cpp11.cpp
@@ -47,9 +47,17 @@ extern "C" SEXP _cppdoubles_cpp_double_rel_diff_vectorised(SEXP x, SEXP y) {
     return cpp11::as_sexp(cpp_double_rel_diff_vectorised(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(y)));
   END_CPP11
 }
+// cpp_doubles.cpp
+SEXP cpp_all_equal(SEXP x, SEXP y, SEXP tolerance, SEXP na_rm);
+extern "C" SEXP _cppdoubles_cpp_all_equal(SEXP x, SEXP y, SEXP tolerance, SEXP na_rm) {
+  BEGIN_CPP11
+    return cpp11::as_sexp(cpp_all_equal(cpp11::as_cpp<cpp11::decay_t<SEXP>>(x), cpp11::as_cpp<cpp11::decay_t<SEXP>>(y), cpp11::as_cpp<cpp11::decay_t<SEXP>>(tolerance), cpp11::as_cpp<cpp11::decay_t<SEXP>>(na_rm)));
+  END_CPP11
+}
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
+    {"_cppdoubles_cpp_all_equal",                  (DL_FUNC) &_cppdoubles_cpp_all_equal,                  4},
     {"_cppdoubles_cpp_double_equal_vectorised",    (DL_FUNC) &_cppdoubles_cpp_double_equal_vectorised,    3},
     {"_cppdoubles_cpp_double_gt_vectorised",       (DL_FUNC) &_cppdoubles_cpp_double_gt_vectorised,       3},
     {"_cppdoubles_cpp_double_gte_vectorised",      (DL_FUNC) &_cppdoubles_cpp_double_gte_vectorised,      3},