+++ src/cpp_doubles.cpp
@@ -5,8 +5,11 @@
 
 #define R_NO_REMAP
 
+// Author: Nick Christofides
+// Date: 20-Feb-2024
+
 // Below are a complete set of C++ functions for comparing doubles
-// mimicing the ==, <=, <, > and >= operators with a tolerance
+// mimicking the ==, <=, <, > and >= operators with a tolerance
 
 // Relative differences are used except when either x or y is very close to zero
 // in which case absolute differences are used
@@ -92,8 +95,7 @@ SEXP cpp_double_equal_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
   R_xlen_t tol_len = Rf_xlength(tolerance);
-  R_xlen_t n = std::fmax(x_len, y_len);
-  n = std::fmax(n, tol_len);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
   if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -103,16 +105,16 @@ SEXP cpp_double_equal_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   double *p_t = REAL(tolerance);
   SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
   int *p_out = LOGICAL(out);
-  R_xlen_t xi;
-  R_xlen_t yi;
-  R_xlen_t ti;
+  R_xlen_t xi, yi, ti;
   for (R_xlen_t i = 0; i < n; ++i) {
     xi = i % x_len;
     yi = i % y_len;
     ti = i % tol_len;
     p_out[i] = cpp_double_equal(p_x[xi], p_y[yi], p_t[ti]);
     // If either is NA, out is NA
-    if ( cpp_double_is_na(p_x[xi]) || cpp_double_is_na(p_y[yi]) ){
+    if (cpp_double_is_na(p_x[xi]) ||
+        cpp_double_is_na(p_y[yi]) ||
+        cpp_double_is_na(p_t[ti])){
       p_out[i] = NA_LOGICAL;
     }
   }
@@ -126,8 +128,7 @@ SEXP cpp_double_gt_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
   R_xlen_t tol_len = Rf_xlength(tolerance);
-  R_xlen_t n = std::fmax(x_len, y_len);
-  n = std::fmax(n, tol_len);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
   if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -137,16 +138,16 @@ SEXP cpp_double_gt_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   double *p_t = REAL(tolerance);
   SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
   int *p_out = LOGICAL(out);
-  R_xlen_t xi;
-  R_xlen_t yi;
-  R_xlen_t ti;
+  R_xlen_t xi, yi, ti;
   for (R_xlen_t i = 0; i < n; ++i) {
     xi = i % x_len;
     yi = i % y_len;
     ti = i % tol_len;
     p_out[i] = cpp_double_gt(p_x[xi], p_y[yi], p_t[ti]);
     // If either is NA, out is NA
-    if ( cpp_double_is_na(p_x[xi]) || cpp_double_is_na(p_y[yi]) ){
+    if (cpp_double_is_na(p_x[xi]) ||
+        cpp_double_is_na(p_y[yi]) ||
+        cpp_double_is_na(p_t[ti])){
       p_out[i] = NA_LOGICAL;
     }
   }
@@ -160,8 +161,7 @@ SEXP cpp_double_gte_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
   R_xlen_t tol_len = Rf_xlength(tolerance);
-  R_xlen_t n = std::fmax(x_len, y_len);
-  n = std::fmax(n, tol_len);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
   if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -171,16 +171,16 @@ SEXP cpp_double_gte_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   double *p_t = REAL(tolerance);
   SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
   int *p_out = LOGICAL(out);
-  R_xlen_t xi;
-  R_xlen_t yi;
-  R_xlen_t ti;
+  R_xlen_t xi, yi, ti;
   for (R_xlen_t i = 0; i < n; ++i) {
     xi = i % x_len;
     yi = i % y_len;
     ti = i % tol_len;
     p_out[i] = cpp_double_gte(p_x[xi], p_y[yi], p_t[ti]);
     // If either is NA, out is NA
-    if ( cpp_double_is_na(p_x[xi]) || cpp_double_is_na(p_y[yi]) ){
+    if (cpp_double_is_na(p_x[xi]) ||
+        cpp_double_is_na(p_y[yi]) ||
+        cpp_double_is_na(p_t[ti])){
       p_out[i] = NA_LOGICAL;
     }
   }
@@ -194,8 +194,7 @@ SEXP cpp_double_lt_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
   R_xlen_t tol_len = Rf_xlength(tolerance);
-  R_xlen_t n = std::fmax(x_len, y_len);
-  n = std::fmax(n, tol_len);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
   if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -205,16 +204,16 @@ SEXP cpp_double_lt_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   double *p_t = REAL(tolerance);
   SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
   int *p_out = LOGICAL(out);
-  R_xlen_t xi;
-  R_xlen_t yi;
-  R_xlen_t ti;
+  R_xlen_t xi, yi, ti;
   for (R_xlen_t i = 0; i < n; ++i) {
     xi = i % x_len;
     yi = i % y_len;
     ti = i % tol_len;
     p_out[i] = cpp_double_lt(p_x[xi], p_y[yi], p_t[ti]);
     // If either is NA, out is NA
-    if ( cpp_double_is_na(p_x[xi]) || cpp_double_is_na(p_y[yi]) ){
+    if (cpp_double_is_na(p_x[xi]) ||
+        cpp_double_is_na(p_y[yi]) ||
+        cpp_double_is_na(p_t[ti])){
       p_out[i] = NA_LOGICAL;
     }
   }
@@ -228,8 +227,7 @@ SEXP cpp_double_lte_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
   R_xlen_t tol_len = Rf_xlength(tolerance);
-  R_xlen_t n = std::fmax(x_len, y_len);
-  n = std::fmax(n, tol_len);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
   if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -239,16 +237,16 @@ SEXP cpp_double_lte_vectorised(SEXP x, SEXP y, SEXP tolerance) {
   double *p_t = REAL(tolerance);
   SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
   int *p_out = LOGICAL(out);
-  R_xlen_t xi;
-  R_xlen_t yi;
-  R_xlen_t ti;
+  R_xlen_t xi, yi, ti;
   for (R_xlen_t i = 0; i < n; ++i) {
     xi = i % x_len;
     yi = i % y_len;
     ti = i % tol_len;
     p_out[i] = cpp_double_lte(p_x[xi], p_y[yi], p_t[ti]);
     // If either is NA, out is NA
-    if ( cpp_double_is_na(p_x[xi]) || cpp_double_is_na(p_y[yi]) ){
+    if (cpp_double_is_na(p_x[xi]) ||
+        cpp_double_is_na(p_y[yi]) ||
+        cpp_double_is_na(p_t[ti])){
       p_out[i] = NA_LOGICAL;
     }
   }
@@ -260,7 +258,7 @@ SEXP cpp_double_lte_vectorised(SEXP x, SEXP y, SEXP tolerance) {
 SEXP cpp_double_rel_diff_vectorised(SEXP x, SEXP y) {
   R_xlen_t x_len = Rf_xlength(x);
   R_xlen_t y_len = Rf_xlength(y);
-  R_xlen_t n = std::fmax(x_len, y_len);
+  R_xlen_t n = std::max(x_len, y_len);
   if (x_len <= 0 || y_len <= 0){
     // Avoid loop if any are length zero vectors
     n = 0;
@@ -275,6 +273,50 @@ SEXP cpp_double_rel_diff_vectorised(SEXP x, SEXP y) {
     xi = i % x_len;
     yi = i % y_len;
     p_out[i] = cpp_double_rel_diff(p_x[xi], p_y[yi]);
+  }
+  Rf_unprotect(1);
+  return out;
+}
+[[cpp11::register]]
+SEXP cpp_all_equal(SEXP x, SEXP y, SEXP tolerance, SEXP na_rm) {
+  if (Rf_length(na_rm) != 1 || !Rf_isLogical(na_rm)){
+    Rf_error("na.rm must be a logical vector of length 1");
+  }
+  bool skip_na = Rf_asLogical(na_rm);
+  bool has_na;
+  R_xlen_t x_len = Rf_xlength(x);
+  R_xlen_t y_len = Rf_xlength(y);
+  R_xlen_t tol_len = Rf_xlength(tolerance);
+  R_xlen_t n = std::max(std::max(x_len, y_len), tol_len);
+  if (x_len <= 0 || y_len <= 0 || tol_len <= 0){
+    // Avoid loop if any are length zero vectors
+    n = 0;
+  }
+  double *p_x = REAL(x);
+  double *p_y = REAL(y);
+  double *p_t = REAL(tolerance);
+  SEXP out = Rf_protect(Rf_allocVector(LGLSXP, 1));
+  LOGICAL(out)[0] = true;
+  R_xlen_t xi, yi, ti;
+  for (R_xlen_t i = 0; i < n; ++i) {
+    xi = i % x_len;
+    yi = i % y_len;
+    ti = i % tol_len;
+    has_na = cpp_double_is_na(p_x[xi]) ||
+      cpp_double_is_na(p_y[yi]) ||
+      cpp_double_is_na(p_t[ti]);
+    if (has_na){
+      if (skip_na){
+        continue;
+      } else {
+        LOGICAL(out)[0] = NA_LOGICAL;
+        break;
+      }
+    }
+    if ( !cpp_double_equal(p_x[xi], p_y[yi], p_t[ti]) ){
+      LOGICAL(out)[0] = false;
+      break;
+    }
   }
   Rf_unprotect(1);
   return out;