+++ NEWS.md
@@ -1,3 +1,10 @@
+# cppdoubles (In development)
+
+* New function `all_equal` as an efficient alternative to `all.equal` for 
+numeric vectors.
+
+* `NA` values are now returned when `tol` is `NA`.
+
 # cppdoubles 0.1.1
 
 * CRAN resubmission.